+++ backend/Dukaan/Dukaan.Infrastructure/Services/OrderNumberService.cs
@@ -1,9 +1,8 @@
+using StackExchange.Redis;
 using Dukaan.Application.Interfaces;
-using Dukaan.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
 using Dukaan.Infrastructure.Data.DbContext;
 using Dukaan.Infrastructure.Services.Interfaces;
-using Microsoft.EntityFrameworkCore;
-using StackExchange.Redis;
 
 namespace Dukaan.Infrastructure.Services;
 
@@ -14,12 +13,11 @@ public class OrderNumberService(
 {
     private const string KeyPrefix = "order_seq:";
 
-    public async Task<(int SequenceNumber, string OrderNumber)> GetNextOrderNumberAsync(
+    public async Task<(long SequenceNumber, string OrderNumber)> GetNextOrderNumberAsync(
         CancellationToken cancellationToken = default)
     {
         var tenantId = tenantProvider.GetTenantId();
-        if (tenantId is null)
-            throw new InvalidOperationException("Tenant ID is not available.");
+        if (tenantId is null) throw new InvalidOperationException("Tenant ID is not available.");
 
         var db = redis.Database;
         var key = $"{KeyPrefix}{tenantId}";
@@ -29,7 +27,7 @@ public class OrderNumberService(
             await SeedFromDatabaseAsync(tenantId.Value, db, key, cancellationToken);
         }
 
-        var sequence = (int)await db.StringIncrementAsync(key);
+        var sequence = await db.StringIncrementAsync(key);
         var orderNumber = $"ORD-{sequence:D6}";
 
         return (sequence, orderNumber);
@@ -40,7 +38,7 @@ public class OrderNumberService(
     {
         var maxSequence = await dbContext.Orders
             .Where(o => o.TenantId == tenantId)
-            .MaxAsync(o => (int?)o.SequenceNumber, cancellationToken) ?? 0;
+            .MaxAsync(o => (long?)o.SequenceNumber, cancellationToken) ?? 0;
 
         await db.StringSetAsync(key, maxSequence);
     }
